Moved tests to FromCallableTest from FromCompletableTest (#5705)
